+++ src/src/WebServer/HTML_wrappers.h
@@ -140,11 +140,11 @@ void addHtmlInt(int32_t int_val);
 void addHtmlInt(uint32_t int_val);
 void addHtmlInt(int64_t int_val);
 void addHtmlInt(uint64_t int_val);
-void addHtmlFloat(const float& value, unsigned int nrDecimals = 2u);
-//void addHtmlFloat_NaN_toNull(const float& value, unsigned int nrDecimals = 2u);
+void addHtmlFloat(const float& value, uint8_t nrDecimals = 2u);
+//void addHtmlFloat_NaN_toNull(const float& value, uint8_t nrDecimals = 2u);
 #if FEATURE_USE_DOUBLE_AS_ESPEASY_RULES_FLOAT_TYPE
-void addHtmlFloat(const double& value, unsigned int nrDecimals = 2u);
-//void addHtmlFloat_NaN_toNull(const double& value, unsigned int nrDecimals = 2u);
+void addHtmlFloat(const double& value, uint8_t nrDecimals = 2u);
+//void addHtmlFloat_NaN_toNull(const double& value, uint8_t nrDecimals = 2u);
 #endif
 
 void addEncodedHtml(const __FlashStringHelper * html);
